Adding base app functionality. Logic for calculating file size needs to be revised as the expected filesize is nowhere near where it should be.

+++ AudioDeliveryManagementSystem/WaveFile.cs
@@ -12,9 +12,9 @@ namespace AudioDeliveryManagementSystem
         public WaveFileStatus Status { get; set; }
         public string FilePath { get; set; }
         public string FileName { get; set; }
-        public long FileSizeInBytes { get; set; }
-        public long ExpectedFileSizeInBytes { get; set; }
-        public long DurationInMillieSecs { get; set; }
+        public decimal FileSizeInBytes { get; set; }
+        public decimal ExpectedFileSizeInBytes { get; set; }
+        public decimal DurationInMillieSecs { get; set; }
         public bool PassedValidation { get; set; }
         public WaveFile(string filePath)
         {
